REFACTORIZADOS controller y service de bienvenida para que sigan el nuevo patrón.
