+++ src/main/java/app/controllers/BienvenidaController.java
@@ -2,37 +2,40 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package app.services;
+
+package app.controllers;
 
 import app.frames.FrameBienvenida;
+import app.services.BienvenidaService;
 import java.util.concurrent.TimeUnit;
 
 /**
  *
- * @author Juan José Molano Franco
+ * @author Juan Jose Molano Franco
  */
-public class ViewBienvenidaService {
-
+public class BienvenidaController {
+    
+    private final BienvenidaService service;
+    
+    public BienvenidaController() {
+        this.service = new BienvenidaService();
+    }
+    
     public String bienvenida() {
+        int seleccion = levantarFrameBienvenida();
+        return service.bienvenida(seleccion);
+    }
+
+    private int levantarFrameBienvenida() {
         FrameBienvenida frame = new FrameBienvenida();
         frame.setVisible(true);
 
         /* Este ciclo revisa cada 250ms si el usuario ya clickeó un botón.
-        En caso de que no espera otros 250ms, pero si sí entonces toma la ruta */
+        En caso de que no espera otros 250ms, pero si sí entonces retorna la selección */
         while (true) {
             int seleccion = frame.getSeleccion();
             switch (seleccion) {
-                case 0:  // Se cierra el frame con la X
-                case 1:  // Cerrar
-                    frame.dispose();
-                    return "salida";
-                case 2:  // Iniciar sesión
-                    frame.dispose();
-                    return "Login_pideNombreUsuario";
-                case 3:  // Iniciar sesión
-                    frame.dispose();
-                    return "Register_pideNombreUsuario";
-                default:  // Valor default, preferiblemente un número negativo
+                case -1:  // Valor default, preferiblemente un número negativo
                     // Detiene el ciclo por 250 ms
                     try {
                         TimeUnit.MILLISECONDS.sleep(250);
@@ -40,6 +43,9 @@ public class ViewBienvenidaService {
                         Thread.currentThread().interrupt();
                     }
                     break;
+                default:
+                    frame.dispose();
+                    return seleccion;
             }
         }
     }
